test: merge red-attacker tests into canonical files and deduplicate
Merge unique adversarial tests from 5 red test files into their canonical
counterparts, dropping 47 tests that duplicated existing coverage or tested
the same code path through different entry points (741 → 694 tests).

Red files removed: codex-jobs-red, red-session-naming, red-coral-agent,
codex-executor-red, cli-detection-red.
