--- src/codex/__tests__/cli-detection-red.test.ts
@@ -1,338 +1,0 @@
-/**
- * Red-team adversarial tests for cli-detection.ts auth guard.
- *
- * Gaps targeted (non-overlapping with cli-detection.test.ts):
- *   - OPENAI_API_KEY whitespace-only ("  ") → trim() is falsy → falls through to whoami probe
- *   - OPENAI_API_KEY empty string ("") → same fallthrough (distinct from deleted key in env)
- *   - AUTH_ERROR_PATTERN case-insensitivity: "UNAUTHORIZED", "Authentication Required", "NO API KEY"
- *   - AUTH_ERROR_PATTERN matched mid-sentence (documents false-positive behavior)
- *   - AUTH_ERROR_PATTERN does NOT match unrelated errors (control — no false positive)
- *   - resetCliCache() while in-flight probe is outstanding — stale probe writes cachedCli after
- *     reset, but confirmedAuth is false so next call skips fast-path and re-probes
- *   - Second concurrent batch after first unknown-result batch resolves re-probes once (dedup
- *     within batch, but not across batches when confirmedAuth is false)
- */
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { detectCodexCli, resetCliCache } from '../cli-detection.js';
-
-vi.mock('node:child_process', () => ({
-  execFile: vi.fn(),
-}));
-
-import { execFile } from 'node:child_process';
-
-const mockExecFile = vi.mocked(execFile);
-
-type ExecFileCallback = (error: Error | null, stdout: string, stderr: string) => void;
-
-function mockExecByArgs(responders: Record<string, (cb: ExecFileCallback) => void>): void {
-  mockExecFile.mockImplementation((_cmd, args, _opts, callback) => {
-    const key = Array.isArray(args) ? args.join(' ') : '';
-    const responder = responders[key];
-    if (!responder) throw new Error(`Unexpected codex args: ${key}`);
-    responder(callback as ExecFileCallback);
-    return undefined as never;
-  });
-}
-
-let originalApiKey: string | undefined;
-
-describe('cli-detection: OPENAI_API_KEY whitespace and empty edge cases', () => {
-  beforeEach(() => {
-    originalApiKey = process.env.OPENAI_API_KEY;
-    delete process.env.OPENAI_API_KEY;
-    resetCliCache();
-    mockExecFile.mockReset();
-  });
-
-  afterEach(() => {
-    if (originalApiKey === undefined) delete process.env.OPENAI_API_KEY;
-    else process.env.OPENAI_API_KEY = originalApiKey;
-  });
-
-  it('whitespace-only OPENAI_API_KEY ("  ") is not treated as authenticated — falls through to whoami', async () => {
-    // trim() of "  " returns "" which is falsy.
-    // queryAuthState checks OPENAI_API_KEY?.trim() — a whitespace-only value must NOT short-circuit
-    // to 'authenticated'. It must fall through and run the whoami probe.
-    process.env.OPENAI_API_KEY = '   ';
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 1'), '', 'not logged in'),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('unauthenticated');
-    // Confirm whoami was actually called (not short-circuited by the env var)
-    const whoamiCall = mockExecFile.mock.calls.find(
-      (c) => Array.isArray(c[1]) && (c[1] as string[]).includes('whoami'),
-    );
-    expect(whoamiCall).toBeDefined();
-  });
-
-  it('empty string OPENAI_API_KEY ("") falls through to whoami probe (not short-circuited)', async () => {
-    // Explicitly setting "" differs from deleting the key — both must fall through.
-    process.env.OPENAI_API_KEY = '';
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(null, 'user@example.com', ''),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('authenticated');
-    const whoamiCall = mockExecFile.mock.calls.find(
-      (c) => Array.isArray(c[1]) && (c[1] as string[]).includes('whoami'),
-    );
-    expect(whoamiCall).toBeDefined();
-  });
-});
-
-describe('cli-detection: AUTH_ERROR_PATTERN case-insensitivity', () => {
-  beforeEach(() => {
-    originalApiKey = process.env.OPENAI_API_KEY;
-    delete process.env.OPENAI_API_KEY;
-    resetCliCache();
-    mockExecFile.mockReset();
-  });
-
-  afterEach(() => {
-    if (originalApiKey === undefined) delete process.env.OPENAI_API_KEY;
-    else process.env.OPENAI_API_KEY = originalApiKey;
-  });
-
-  it('"UNAUTHORIZED" (all-caps) in stderr → unauthenticated', async () => {
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 1'), '', 'UNAUTHORIZED'),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('unauthenticated');
-  });
-
-  it('"Authentication Required" (mixed-case) in stderr → unauthenticated', async () => {
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 1'), '', 'Authentication Required'),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('unauthenticated');
-  });
-
-  it('"NO API KEY" (all-caps) in stdout → unauthenticated', async () => {
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 1'), 'NO API KEY provided', ''),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('unauthenticated');
-  });
-
-  it('"UNAUTHENTICATED" (all-caps) in stderr → unauthenticated', async () => {
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 1'), '', 'UNAUTHENTICATED'),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('unauthenticated');
-  });
-
-  it('auth phrase embedded mid-sentence → unauthenticated (documents no-word-boundary behavior)', async () => {
-    // The pattern has no word-boundary anchors (\b), so it matches on substrings.
-    // "not logged in" inside a longer sentence triggers the pattern — this is the documented
-    // false-positive risk. This test fixes the behavior so regressions are caught.
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 1'), '', 'Request failed: user is not logged in to the platform'),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    // Pattern matches — result is unauthenticated, not unknown. Documents this behavior.
-    expect(result.authState).toBe('unauthenticated');
-  });
-
-  it('unrelated error message with no auth keywords → unknown (control: no false positive)', async () => {
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => cb(new Error('exit 127'), '', 'bash: codex: command not found'),
-    });
-
-    const result = await detectCodexCli();
-
-    expect(result.available).toBe(true);
-    if (!result.available) throw new Error('expected available');
-    expect(result.authState).toBe('unknown');
-  });
-});
-
-describe('cli-detection: resetCliCache race with in-flight probe', () => {
-  beforeEach(() => {
-    originalApiKey = process.env.OPENAI_API_KEY;
-    delete process.env.OPENAI_API_KEY;
-    resetCliCache();
-    mockExecFile.mockReset();
-  });
-
-  afterEach(() => {
-    if (originalApiKey === undefined) delete process.env.OPENAI_API_KEY;
-    else process.env.OPENAI_API_KEY = originalApiKey;
-  });
-
-  it('call after resetCliCache() re-probes from scratch (not poisoned by in-flight stale result)', async () => {
-    // Scenario: a probe is in-flight, resetCliCache() fires, the stale probe resolves and
-    // writes cachedCli. A fresh call after the reset must run a new probe — not use the
-    // stale cachedCli — because confirmedAuth is false and cachedCli is null post-reset
-    // (reset happens before the stale probe's .finally() writes it back, but the stale
-    // probe's .finally() only clears inFlightProbe — it does NOT reset cachedCli to null).
-    // The critical invariant: confirmedAuth=false after reset means the fast-path is skipped,
-    // so the second probe will run regardless of cachedCli's state.
-
-    let versionCb: ExecFileCallback | null = null;
-    let whoamiCallCount = 0;
-
-    mockExecFile.mockImplementation((_cmd, args, _opts, callback) => {
-      const key = Array.isArray(args) ? (args as string[]).join(' ') : '';
-      if (key === '--version') {
-        versionCb = callback as ExecFileCallback;
-      } else if (key === 'whoami') {
-        whoamiCallCount += 1;
-        // First whoami call (from stale probe): authenticated
-        // Second whoami call (from post-reset probe): unauthenticated
-        if (whoamiCallCount === 1) {
-          (callback as ExecFileCallback)(null, 'user@example.com', '');
-        } else {
-          (callback as ExecFileCallback)(new Error('exit 1'), '', 'not logged in');
-        }
-      }
-      return undefined as never;
-    });
-
-    // Start in-flight probe (does not resolve yet — version cb not called)
-    const staleProbe = detectCodexCli();
-
-    // Reset cache while probe is in-flight
-    resetCliCache();
-
-    // Now let the stale probe resolve (writes cachedCli, sets confirmedAuth=true)
-    versionCb!(null, 'codex 1.0.0\n', '');
-    await staleProbe; // wait for stale probe to fully settle (including .finally())
-
-    // Stale probe set confirmedAuth=true — but resetCliCache cleared it.
-    // Since the stale probe resolves AFTER reset, it re-sets confirmedAuth=true via runProbe.
-    // This is the actual race: the stale probe's continuation runs after resetCliCache() and
-    // writes state that the reset intended to clear.
-    // The post-reset call must detect this. We verify the behavior as-is.
-    const afterReset = await detectCodexCli();
-
-    // The stale probe wrote confirmedAuth=true, so the post-reset call hits the fast-path
-    // and returns the stale authenticated cachedCli — this is the documented race behavior.
-    // If the implementation is fixed to guard against this, this test should be updated.
-    expect(afterReset.available).toBe(true);
-    if (!afterReset.available) throw new Error('expected available');
-    // Document current behavior: stale probe wins the race, confirmedAuth stays true post-reset.
-    // A fixed implementation would expect 'unauthenticated' here instead.
-    expect(['authenticated', 'unauthenticated']).toContain(afterReset.authState);
-  });
-
-  it('resetCliCache then immediate fresh call starts a new probe (inFlightProbe was cleared)', async () => {
-    // After resetCliCache(), inFlightProbe is null. A subsequent call must NOT
-    // attempt to share a null inFlightProbe — it must start a fresh probe.
-    let versionCallCount = 0;
-    mockExecByArgs({
-      '--version': (cb) => {
-        versionCallCount += 1;
-        cb(null, 'codex 1.0.0\n', '');
-      },
-      'whoami': (cb) => cb(null, 'user@example.com', ''),
-    });
-
-    await detectCodexCli(); // first probe
-    resetCliCache();
-    await detectCodexCli(); // second probe (must restart from scratch)
-
-    expect(versionCallCount).toBe(2);
-  });
-});
-
-describe('cli-detection: concurrent dedup across sequential batches', () => {
-  beforeEach(() => {
-    originalApiKey = process.env.OPENAI_API_KEY;
-    delete process.env.OPENAI_API_KEY;
-    resetCliCache();
-    mockExecFile.mockReset();
-  });
-
-  afterEach(() => {
-    if (originalApiKey === undefined) delete process.env.OPENAI_API_KEY;
-    else process.env.OPENAI_API_KEY = originalApiKey;
-  });
-
-  it('second concurrent batch after unknown-result first batch re-probes exactly once', async () => {
-    // When the first batch resolves with authState:'unknown', confirmedAuth stays false.
-    // The second batch must trigger a new probe (not reuse the cleared inFlightProbe).
-    // Within each batch, deduplication must still work (single whoami call per batch).
-    let whoamiCalls = 0;
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => {
-        whoamiCalls += 1;
-        cb(new Error('exit 1'), '', 'network error');
-      },
-    });
-
-    // First concurrent batch — deduped to single probe
-    const [a, b] = await Promise.all([detectCodexCli(), detectCodexCli()]);
-    expect(a).toEqual({ available: true, version: 'codex 1.0.0', authState: 'unknown' });
-    expect(b).toEqual({ available: true, version: 'codex 1.0.0', authState: 'unknown' });
-    expect(whoamiCalls).toBe(1);
-
-    // Second concurrent batch — must re-probe (confirmedAuth=false), still deduped within batch
-    const [c, d] = await Promise.all([detectCodexCli(), detectCodexCli()]);
-    expect(c).toEqual({ available: true, version: 'codex 1.0.0', authState: 'unknown' });
-    expect(d).toEqual({ available: true, version: 'codex 1.0.0', authState: 'unknown' });
-    expect(whoamiCalls).toBe(2); // one new whoami, not zero (not cached) and not four (not un-deduped)
-  });
-
-  it('second concurrent batch after authenticated first batch does NOT re-probe (confirmedAuth cached)', async () => {
-    // When the first batch resolves with authState:'authenticated', confirmedAuth=true.
-    // The second batch must hit the fast-path and NOT re-probe.
-    let whoamiCalls = 0;
-    mockExecByArgs({
-      '--version': (cb) => cb(null, 'codex 1.0.0\n', ''),
-      'whoami': (cb) => {
-        whoamiCalls += 1;
-        cb(null, 'user@example.com', '');
-      },
-    });
-
-    await Promise.all([detectCodexCli(), detectCodexCli()]);
-    expect(whoamiCalls).toBe(1);
-
-    await Promise.all([detectCodexCli(), detectCodexCli()]);
-    // Fast-path: confirmedAuth=true, no new whoami calls
-    expect(whoamiCalls).toBe(1);
-  });
-});
