--- src/codex/__tests__/red-coral-agent.test.ts
@@ -1,547 +1,0 @@
-/**
- * Red-team adversarial tests for the coral:* dynamic agent op feature.
- *
- * STAGING FILE — place at src/codex/__tests__/red-coral-agent.test.ts to run.
- * All imports below use paths relative to src/codex/__tests__/.
- *
- * Targeted modules:
- *   src/codex/schemas.ts         — coralAgentSchema regex boundary cases
- *   src/codex/server-handlers.ts — resolveAgentPrompt, handleCoralAgent, routing
- *   src/codex/codex-executor.ts  — ensureMultiAgent fail-open (AC6: no flag on write error)
- *
- * Coverage gaps targeted (non-overlapping with existing tests):
- *
- *  SCHEMA (coralAgentSchema):
- *   1. Single-char agent name `coral:a` accepted — minimum valid length not tested
- *   2. Uppercase in name rejected (`coral:Scanner`) — existing test only covers non-coral prefix
- *   3. Underscore rejected (`coral:scanner_two`) — stricter than identPattern; not tested
- *   4. Name starting with hyphen rejected (`coral:-scanner`) — start char must be [a-z0-9]
- *   5. Double-colon (`coral::scanner`) rejected — `:` not in [a-z0-9-]
- *   6. All-digits name (`coral:123`) accepted — digits valid in [a-z0-9]
- *   7. Double-traversal `coral:../../tmp/x` rejected (vs single `coral:../x` in existing tests)
- *   8. All optional fields accepted alongside required op+prompt
- *
- *  HANDLER - resolveAgentPrompt:
- *   9. Returns file contents exactly as-is including YAML frontmatter (AC14: no stripping)
- *  10. typeof discrimination: missing agent → isError McpResult, NOT thrown (no double-wrapping)
- *  11. File content is not trimmed or modified
- *
- *  HANDLER - handleCoralAgent session path (ordering invariant):
- *  12. Session provided NOT in mgr → sessionNotFoundError before preflight
- *      (session lookup runs before CLI check — detectCodexCli must NOT be called)
- *  13. Session found + CLI unauthenticated → auth error fires after session lookup
- *
- *  HANDLER - session name and prompt construction:
- *  14. Explicit `name` field overrides the generated `agentName-timestamp` label
- *  15. Without explicit name, session_name follows `agentName-\d+` pattern
- *  16. Separator is exactly `\n\n---\n\n` — double-newline on each side
- *  17. Agent file containing `---` in body does not confuse the separator
- *
- *  HANDLER - session resume (AC7):
- *  18. coral:scanner with session= dispatches via executeResume, not executeOneShot
- *  19. Augmented prompt is sent to executeResume with agent content prepended
- *  20. working_directory falls back to session entry's cwd when not provided
- *
- *  HANDLER - tool schema (AC5/AC11):
- *  21. op description documents coral:* support (not just enum absence)
- *
- *  HANDLER - Zod validation error routing:
- *  22. Empty prompt string fails Zod, not file lookup
- *  23. Invalid reasoning_effort fails Zod, not file lookup
- *  24. Missing prompt → isError with Zod content, never unknown_op
- *
- *  HANDLER - path boundary check (defense-in-depth):
- *  25. Valid agent name reaches file-read path (not path-traversal error)
- *  26. Non-existent agent → file-not-found error (not path-traversal error)
- *
- *  EXECUTOR - ensureMultiAgent (AC6):
- *  27. _test.setMultiAgentEnsured API exists and allows resetting (proves flag is not permanent)
- */
-
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import {
-  mkdtempSync,
-  mkdirSync,
-  rmSync,
-  writeFileSync,
-  readFileSync,
-} from 'node:fs';
-import { join } from 'node:path';
-import { ZodError } from 'zod';
-
-import { coralAgentSchema } from '../schemas.js';
-
-vi.mock('../codex-executor.js', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../codex-executor.js')>();
-  return {
-    ...actual,
-    executeOneShot: vi.fn(),
-    executeResume: vi.fn(),
-    executeFork: vi.fn(),
-  };
-});
-
-vi.mock('../cli-detection.js', () => ({
-  detectCodexCli: vi.fn(),
-}));
-
-vi.mock('../progress.js', () => ({
-  createSessionDir: vi.fn(() => ({
-    id: '11111111-2222-4333-8444-555555555555',
-    dir: '/tmp/coral-sessions/11111111-2222-4333-8444-555555555555',
-  })),
-  writeSessionResult: vi.fn(),
-  writeSessionError: vi.fn(),
-  readSessionStatus: vi.fn(() => ({ status: 'running' })),
-  resolveSessionDir: vi.fn((id: string) => `/tmp/coral-sessions/${id}`),
-  SESSIONS_DIR: '/tmp/coral-sessions',
-  PROGRESS_FILE: 'progress.jsonl',
-  extractProgressMessage: vi.fn(),
-  appendProgressEvent: vi.fn(),
-  formatElapsed: vi.fn(() => ''),
-}));
-
-import { executeOneShot, executeResume, _test as executorTest } from '../codex-executor.js';
-import { detectCodexCli } from '../cli-detection.js';
-import {
-  handleToolCall,
-  activeSessions,
-  tools,
-  _test as handlerTest,
-} from '../server-handlers.js';
-import { SessionManager } from '../session-manager.js';
-import type { CodexExecResult } from '../../types.js';
-
-function makeExecResult(overrides: Partial<CodexExecResult> = {}): CodexExecResult {
-  return {
-    response: 'ok',
-    sessionId: 'thread-red-001',
-    model: 'o4-mini',
-    durationMs: 100,
-    exitCode: 0,
-    errors: [],
-    warnings: [],
-    aborted: false,
-    ...overrides,
-  };
-}
-
-let tmpDir = '';
-let mgr: SessionManager;
-const defaultPluginRoot = process.cwd();
-
-beforeEach(() => {
-  tmpDir = mkdtempSync(join('/tmp', 'coral-red-test-'));
-  mkdirSync(join(tmpDir, 'agents'), { recursive: true });
-  mkdirSync(join(tmpDir, 'workspace'), { recursive: true });
-  writeFileSync(join(tmpDir, 'agents', 'scanner.md'), '# Scanner\nDo analysis.\n');
-  handlerTest.setPluginRoot(tmpDir);
-  mgr = new SessionManager(join(tmpDir, 'workspace'));
-  activeSessions.clear();
-  vi.mocked(detectCodexCli).mockResolvedValue({
-    available: true,
-    version: 'codex 1.0.0',
-    authState: 'authenticated',
-  });
-  vi.mocked(executeOneShot).mockResolvedValue(makeExecResult());
-  vi.mocked(executeResume).mockResolvedValue(makeExecResult());
-});
-
-afterEach(() => {
-  activeSessions.clear();
-  vi.clearAllMocks();
-  handlerTest.setPluginRoot(defaultPluginRoot);
-  rmSync(tmpDir, { recursive: true, force: true });
-});
-
-// ─── 1-8: Schema boundary cases ───────────────────────────────────────────────
-
-describe('coralAgentSchema: regex boundary cases', () => {
-  it('accepts single-char agent name coral:a (minimum valid length)', () => {
-    // Regex: /^coral:[a-z0-9][a-z0-9-]*$/ — the trailing * allows zero chars after first,
-    // so `coral:a` (one char after colon) is valid.
-    expect(() => coralAgentSchema.parse({ op: 'coral:a', prompt: 'go' })).not.toThrow();
-  });
-
-  it('accepts all-digits agent name coral:123', () => {
-    // Digits are in [a-z0-9] — all-numeric names are valid under the regex.
-    expect(() => coralAgentSchema.parse({ op: 'coral:123', prompt: 'go' })).not.toThrow();
-  });
-
-  it('rejects uppercase letter in agent name coral:Scanner', () => {
-    // Stricter than identPattern: uppercase is not in [a-z0-9-].
-    expect(() => coralAgentSchema.parse({ op: 'coral:Scanner', prompt: 'go' })).toThrow(ZodError);
-  });
-
-  it('rejects all-uppercase agent name coral:SCANNER', () => {
-    expect(() => coralAgentSchema.parse({ op: 'coral:SCANNER', prompt: 'go' })).toThrow(ZodError);
-  });
-
-  it('rejects underscore in agent name coral:scanner_two', () => {
-    // Underscore is not in [a-z0-9-] — plan explicitly states kebab-case only.
-    expect(() => coralAgentSchema.parse({ op: 'coral:scanner_two', prompt: 'go' })).toThrow(ZodError);
-  });
-
-  it('rejects agent name starting with hyphen coral:-scanner', () => {
-    // First char must be [a-z0-9]; hyphen is only valid after the first character.
-    expect(() => coralAgentSchema.parse({ op: 'coral:-scanner', prompt: 'go' })).toThrow(ZodError);
-  });
-
-  it('rejects double-colon prefix coral::scanner', () => {
-    // The agent name starts with `:`; not in [a-z0-9].
-    expect(() => coralAgentSchema.parse({ op: 'coral::scanner', prompt: 'go' })).toThrow(ZodError);
-  });
-
-  it('rejects double-traversal coral:../../tmp/x (dots blocked by regex)', () => {
-    // Existing tests cover single `coral:../x`; this tests the double form.
-    // Both blocked: `.` is not in [a-z0-9-].
-    expect(() =>
-      coralAgentSchema.parse({ op: 'coral:../../tmp/x', prompt: 'go' }),
-    ).toThrow(ZodError);
-  });
-
-  it('accepts all optional fields alongside required op and prompt', () => {
-    const result = coralAgentSchema.parse({
-      op: 'coral:scanner',
-      prompt: 'analyze',
-      model: 'o4-mini',
-      working_directory: '/tmp',
-      reasoning_effort: 'high',
-      bypass: true,
-    });
-    expect(result).toMatchObject({
-      op: 'coral:scanner',
-      prompt: 'analyze',
-      model: 'o4-mini',
-      working_directory: '/tmp',
-      reasoning_effort: 'high',
-      bypass: true,
-    });
-  });
-});
-
-// ─── 9-11: resolveAgentPrompt file content and return type ────────────────────
-
-describe('resolveAgentPrompt: file content and return type discrimination', () => {
-  it('returns file contents as-is including YAML frontmatter — no stripping (AC14)', async () => {
-    // Frontmatter is plain text from the perspective of the prepend logic.
-    // No YAML parsing or stripping should occur.
-    const frontmatterContent = [
-      '---',
-      'title: Test Agent',
-      'model: o4-mini',
-      '---',
-      '',
-      '# Test Agent',
-      'You are a test agent.',
-    ].join('\n') + '\n';
-
-    writeFileSync(join(tmpDir, 'agents', 'frontmatter-agent.md'), frontmatterContent);
-
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:frontmatter-agent', prompt: 'go' },
-      mgr,
-    );
-    expect(result.isError).toBe(false);
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    const calledPrompt = vi.mocked(executeOneShot).mock.calls[0]?.[0];
-    // The frontmatter block must appear unchanged in the augmented prompt
-    expect(calledPrompt).toContain('---\ntitle: Test Agent\nmodel: o4-mini\n---');
-    // User prompt must appear after the separator
-    expect(calledPrompt).toContain('\n\n---\n\ngo');
-  });
-
-  it('missing agent → isError with agent-not-found message, no Error: prefix doubling', async () => {
-    // resolveAgentPrompt returns McpResult directly; handleCoralAgent returns it as-is.
-    // handleToolCall must not wrap it in an additional "Error: " prefix.
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:does-not-exist-xyz', prompt: 'test' },
-      mgr,
-    );
-    expect(result.isError).toBe(true);
-    expect(result.content[0].text).toContain('Agent file not found: agents/does-not-exist-xyz.md');
-    // Must NOT be double-wrapped ("Error: Error: ...")
-    expect(result.content[0].text).not.toMatch(/^Error: Error:/);
-  });
-
-  it('agent file content is not trimmed — leading spaces and trailing newlines preserved', async () => {
-    const exactContent = '  leading space\n\ntrailing newline\n';
-    writeFileSync(join(tmpDir, 'agents', 'exact-content.md'), exactContent);
-
-    await handleToolCall('codex', { op: 'coral:exact-content', prompt: 'x' }, mgr);
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    const calledPrompt = vi.mocked(executeOneShot).mock.calls[0]?.[0];
-    expect(calledPrompt).toContain('  leading space\n\ntrailing newline\n');
-  });
-});
-
-// ─── 12-13: Session path ordering — session lookup before CLI preflight ────────
-
-describe('handleCoralAgent: session lookup ordering before CLI preflight', () => {
-  it('session provided NOT in mgr → sessionNotFoundError; detectCodexCli NOT called', async () => {
-    // Implementation order in handleCoralAgent (session branch):
-    //   1. mgr.get(session)  ← must run first
-    //   2. preflightCliCheck ← only if session found
-    // If session missing, return sessionNotFoundError without touching CLI.
-    vi.mocked(detectCodexCli).mockResolvedValue({ available: false, error: 'CLI not installed' });
-
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', session: 'nonexistent-session', prompt: 'hi' },
-      mgr,
-    );
-
-    expect(result.isError).toBe(true);
-    expect(result.content[0].text).toContain('nonexistent-session');
-    expect(result.content[0].text).not.toContain('CLI not installed');
-    expect(detectCodexCli).not.toHaveBeenCalled();
-  });
-
-  it('session found + CLI unauthenticated → auth error fires after session lookup', async () => {
-    // When session IS found, preflight runs next and auth error is returned.
-    const sessionId = '12345678-1234-4234-8234-123456789abc';
-    mgr.register(sessionId, 'test-session', 'thread-auth-test', 'o4-mini', '/workspace');
-    vi.mocked(detectCodexCli).mockResolvedValue({
-      available: true,
-      version: 'codex 1.0.0',
-      authState: 'unauthenticated',
-      authError: 'Run codex login',
-    });
-
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', session: sessionId, prompt: 'hi' },
-      mgr,
-    );
-
-    expect(result.isError).toBe(true);
-    expect(result.content[0].text).toContain('Run codex login');
-  });
-});
-
-// ─── 14-17: Session name generation and augmented prompt construction ──────────
-
-describe('handleCoralAgent: session name and prompt construction', () => {
-  it('explicit name field overrides generated agentName-timestamp label', async () => {
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', prompt: 'scan this', name: 'my-custom-session' },
-      mgr,
-    );
-    expect(result.isError).toBe(false);
-    const data = JSON.parse(result.content[0].text);
-    expect(data.session_name).toBe('my-custom-session');
-  });
-
-  it('without explicit name, session_name follows agentName-timestamp pattern', async () => {
-    const before = Date.now();
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', prompt: 'scan this' },
-      mgr,
-    );
-    const after = Date.now();
-
-    expect(result.isError).toBe(false);
-    const data = JSON.parse(result.content[0].text);
-    // Must start with the agent name extracted from op (op.slice(6) = 'scanner')
-    expect(data.session_name).toMatch(/^scanner-\d+$/);
-    const ts = parseInt(data.session_name.replace('scanner-', ''), 10);
-    expect(ts).toBeGreaterThanOrEqual(before);
-    expect(ts).toBeLessThanOrEqual(after);
-  });
-
-  it('separator between agent content and user prompt is exactly \\n\\n---\\n\\n', async () => {
-    const agentContent = '# Agent\nContent here.';
-    writeFileSync(join(tmpDir, 'agents', 'sep-test.md'), agentContent);
-
-    await handleToolCall('codex', { op: 'coral:sep-test', prompt: 'user prompt' }, mgr);
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    const calledPrompt = vi.mocked(executeOneShot).mock.calls[0]?.[0];
-    // Implementation: `${agentContent}\n\n---\n\n${input.prompt}`
-    expect(calledPrompt).toContain(`${agentContent}\n\n---\n\nuser prompt`);
-  });
-
-  it('agent file body containing --- does not confuse the separator', async () => {
-    // Frontmatter or section dividers in the agent file must not merge with the appended separator.
-    const agentContent = '# Agent\n---\nSection after divider\n';
-    writeFileSync(join(tmpDir, 'agents', 'has-separator.md'), agentContent);
-
-    await handleToolCall('codex', { op: 'coral:has-separator', prompt: 'user prompt' }, mgr);
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    const calledPrompt = vi.mocked(executeOneShot).mock.calls[0]?.[0];
-    // Both the agent's own --- and the new separator must appear in sequence
-    expect(calledPrompt).toContain('# Agent\n---\nSection after divider\n\n\n---\n\nuser prompt');
-  });
-});
-
-// ─── 18-20: Session continuity — coral:* with session= (AC7) ──────────────────
-
-describe('coral:* with session field resumes existing session (AC7)', () => {
-  it('coral:scanner with session= dispatches via executeResume not executeOneShot', async () => {
-    const sessionId = '11111111-1111-4111-8111-111111111111';
-    mgr.register(sessionId, 'existing-session', 'thread-resume-001', 'o4-mini', '/workspace');
-
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', session: sessionId, prompt: 'follow up' },
-      mgr,
-    );
-    expect(result.isError).toBe(false);
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    expect(executeResume).toHaveBeenCalledTimes(1);
-    expect(executeOneShot).not.toHaveBeenCalled();
-  });
-
-  it('coral:scanner session resume: augmented prompt is sent to executeResume', async () => {
-    const sessionId = '22222222-2222-4222-8222-222222222222';
-    mgr.register(sessionId, 'resume-session', 'thread-resume-002', 'o4-mini', '/workspace');
-    const agentContent = readFileSync(join(tmpDir, 'agents', 'scanner.md'), 'utf-8');
-
-    await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', session: sessionId, prompt: 'analyze again' },
-      mgr,
-    );
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    expect(executeResume).toHaveBeenCalledTimes(1);
-    const calledPrompt = vi.mocked(executeResume).mock.calls[0]?.[1];
-    expect(calledPrompt).toContain(agentContent);
-    expect(calledPrompt).toContain('\n\n---\n\nanalyze again');
-  });
-
-  it('coral:scanner session resume: working_directory falls back to session entry cwd', async () => {
-    const sessionId = '33333333-3333-4333-8333-333333333333';
-    mgr.register(sessionId, 'cwd-session', 'thread-cwd-001', 'o4-mini', '/project/root');
-
-    await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', session: sessionId, prompt: 'scan' },
-      mgr,
-    );
-
-    await new Promise((resolve) => setTimeout(resolve, 30));
-
-    // working_directory not provided in input — must use session entry's workingDirectory
-    const calledCwd = vi.mocked(executeResume).mock.calls[0]?.[3];
-    expect(calledCwd).toBe('/project/root');
-  });
-});
-
-// ─── 21: tools inputSchema documents coral:* ─────────────────────────────────
-
-describe('tools[0].inputSchema: op description documents coral:* (AC11)', () => {
-  it('op description contains the coral: prefix pattern for agent delegation', () => {
-    const opProp = tools[0].inputSchema.properties.op as { description?: string; enum?: unknown[] };
-    expect(opProp.description).toContain('coral:');
-  });
-
-  it('op description provides a concrete coral: example (e.g. coral:scanner)', () => {
-    const opProp = tools[0].inputSchema.properties.op as { description?: string };
-    // Must be actionable documentation, not just "coral:*" abstract notation
-    expect(opProp.description).toMatch(/coral:[a-z]/);
-  });
-
-  it('op field has no enum (existing test re-confirmed for defense-in-depth)', () => {
-    const opProp = tools[0].inputSchema.properties.op as { enum?: unknown[] };
-    expect(opProp.enum).toBeUndefined();
-  });
-});
-
-// ─── 22-24: Zod validation error routing for coral:* ─────────────────────────
-
-describe('coral:* Zod validation: error shape and routing', () => {
-  it('empty prompt string fails Zod validation (not agent file lookup)', async () => {
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', prompt: '' },
-      mgr,
-    );
-    expect(result.isError).toBe(true);
-    // Error must come from promptSchema (min(1)), not from file resolution
-    expect(result.content[0].text).not.toContain('Agent file not found');
-  });
-
-  it('invalid reasoning_effort value fails Zod validation', async () => {
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', prompt: 'go', reasoning_effort: 'ultra-high' } as unknown as Record<string, unknown>,
-      mgr,
-    );
-    expect(result.isError).toBe(true);
-    expect(result.content[0].text).not.toContain('Agent file not found');
-  });
-
-  it('missing prompt → isError Zod content, never unknown_op response', async () => {
-    // coral: prefix intercepts before discriminated union; Zod validates coralAgentSchema.
-    // Must NOT fall through to unknown_op even though coralAgentSchema.safeParse fails.
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner' /* no prompt */ },
-      mgr,
-    );
-    expect(result.isError).toBe(true);
-    const text = result.content[0].text;
-    expect(text).not.toContain('unknown_op');
-    expect(text).not.toMatch(/\{"error":"unknown_op"/);
-  });
-});
-
-// ─── 25-26: Path boundary check — two distinct error paths ───────────────────
-
-describe('resolveAgentPrompt: path-traversal vs file-not-found are distinct errors', () => {
-  it('valid name for existing agent: no path-traversal error emitted', async () => {
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:scanner', prompt: 'test' },
-      mgr,
-    );
-    expect(result.isError).toBe(false);
-    expect(result.content[0].text).not.toContain('Invalid agent name');
-  });
-
-  it('valid-regex name with no file → file-not-found error, not path-traversal error', async () => {
-    // A kebab-case name that passes the regex but has no corresponding .md file.
-    // Must reach the readFileSync catch, NOT the startsWith boundary check.
-    const result = await handleToolCall(
-      'codex',
-      { op: 'coral:ghost-agent', prompt: 'test' },
-      mgr,
-    );
-    expect(result.isError).toBe(true);
-    expect(result.content[0].text).toContain('Agent file not found: agents/ghost-agent.md');
-    // Must NOT produce the path-traversal error message
-    expect(result.content[0].text).not.toContain('Invalid agent name');
-  });
-});
-
-// ─── 27: ensureMultiAgent flag reset API (AC6 structural proof) ───────────────
-
-describe('ensureMultiAgent: _test.setMultiAgentEnsured proves flag is resettable (AC6)', () => {
-  it('_test.setMultiAgentEnsured(false) exists and allows retry on next exec call', () => {
-    // AC6: "ensureMultiAgent does not set flag on failed write"
-    // Structural proof: the _test.setMultiAgentEnsured API allows resetting the flag,
-    // confirming the flag is a regular boolean (not permanently poisoned on failure).
-    // The catch block in ensureMultiAgent does NOT set multiAgentEnsured = true,
-    // which is enforced by the fact that setMultiAgentEnsured(false) can re-enable retries.
-    expect(typeof executorTest.setMultiAgentEnsured).toBe('function');
-    expect(() => executorTest.setMultiAgentEnsured(false)).not.toThrow();
-    // After reset to false, the next executeCodex call will attempt ensureMultiAgent again.
-    // This is the retry behavior required by AC6.
-    expect(() => executorTest.setMultiAgentEnsured(true)).not.toThrow();
-  });
-});
